Sync menu: fix Mass Queue's Enrich commands
On the administration interface's sync menu, selecting
one of the "Enrich" entries on the "Mass Queue" menu
did not actually do an enrich. (But selecting "Sync" did.)

There was an inconsistency between the view and controller
about the internal naming of the enrich task ("full_enrich" versus
"enrich"). It has now been consistently renamed "enrich".

(This naming is now also consistent with the "sync"/"fast_sync"
naming.)