Vocabs CMS: CC-1518 Make version modal "Cancel" button work
When using the vocabs CMS version modal dialog to edit an existing
version, any changes were immediately reflected in the base window,
and clicking the Cancel button did not undo them.

Now, when editing an existing version, operate on a copy of
the version's model, and only apply that if/when the Save button
is clicked.