Vocabs CMS: CC-1518 Make related entity modal "Cancel" button work
When using the vocabs CMS related entity modal dialog to edit an existing
related entity, any changes were immediately reflected in the base window,
and clicking the Cancel button did not undo them.

Now, when editing an existing related entity, operate on a copy of
the related entity's model, and only apply that if/when the Save button
is clicked.