Vocabs CMS: CC-1261 Ensure exactly one pretty name per affiliation
Deal with the number of full names for an affiliation not otherwise being
exactly one (i.e., either 0, in which case, use the affiliation
name as the full name, or greater than 1, in which case, just
pick one).

Also, include both full name and short name in the pretty name.