Vocabs CMS: CC-1261 Use affiliations as keys of affiliationsNames
The controller takes the return value of affiliations() and
applies array_unique() to it.  So the result could then be
"out of sync" with the return value of affiliationsNames().
So change the return value of affiliationsNames() so that
the keys are affiliations, not integers. This ensures
lookup of values in the return value of affiliationsNames()
will be reliable.