CC-1450: Updated/customized Upload.php with mod_mime_fix support
Updated system/libraries/Upload.php from CodeIgniter on GitHub.
It doesn't work "out of the box" because it relies on an extra
function in system/core/Common.php, which we're _not_ also
updating. So I did the minimum modifications to make it
work. Search for "ANDS" to find the customizations.