Vocabs: Add modified-ui-bootstrap-tpls to Gruntfile
This is a completion of commit 1ac6eb731ff8289bfb0bbfaddcc461ce5b34e05f
from 2015-07-15, which added modified-ui-bootstrap-tpls-0.10.0.js
to applications/portal/vocabs/views/includes/scripts.blade.php
but which did not add it to applications/portal/Gruntfile.js.

This commit adds this JS to Gruntfile.js and include the updated lib.js.

Symptom was broken vocabs datepicker behaviour that worked in
development mode but not in production mode.