Vocabs: CC-1268: Type-specific error message for access point
"Add a new version" modal dialog:
For the validation message below the "URI"/"Upload File" field,
only show an error message at all if an Access Point Type has
been selected. If it has, and there is an error, show an error
that is type-specific. Previously, it only ever said
"URI is required". Now, if tye type is "File", it will say
"File is required".