Vocabs CMS: CC-1267 Don't require clicking Add button in most cases
Many -- but not all -- of the existing CMS "Add" buttons required you to click
them _after_ entering data.

Now, if clicking an "Add" button is required, it is clicked
_before_ entering data.

The one remaining exception to this is the access point section
of the version modal. There, you must still click the "Add" button
_after_ filling in the fields.