TR DCI citation updater: cope with MySQL's notion of an "update"
When updating the cited column of dbs_portal's record_stats table,
we attempt an UPDATE first, and then fall back to do an INSERT
if the row is not already present. To test if the UPDATE failed,
we look at the value of cur.rowcount.

But cur.rowcount is zero not only if the row is missing, but if
the row is present, but the update had no effect, e.g., if we
changed the value "1" to "1".

So now, when opening the connection, specify a connection property
that makes such an update count as an update.