Vocabs: Remove fragment from end of My Vocabs Login URL
In the header of the vocabs portal, the link to "My Vocabs Login"
ended with "#/", despite this already being in the fragment part
of the URL. If this is actually needed, the hash needs to be
URL encoded.

But I suspect it is not needed, so I have removed it.