Vocab widget: comment out CSS that goes too far
Vocab widget CSS includes a section that has styling that applies to
all forms. The beginning of that section has the comment
"Demonstration styles (not required for normal operation)".
So it has now been commented out.
If these styles are required in future, add some more selectors
so that the styles don't "leak" everywhere.