Link checkers: ping database before saving results to database
It may be that a long time passes between getting the
links to check, and writing the results of checking back
to the database. And so it can happen that there is a timeout
in the database connection, and writing the results then
fails. So before calling the method to write results,
ping the database, forcing a reconnect if there has been
a timeout.