Vocabs: Specify correct description field in Solr search
The Solr query did not specify the description field with
correct syntax. (It looked like an attempt to do a type
of fuzzy searching, but that is not supported with that
syntax. So the description field was not being searched
at all.) Now, specify the description field so that it
does in fact get searched.