Vocabs CMS: don't try to "fix" language when pre-populating from PP
The values of the $scope.vocab.language array are the
short codes (e.g., "en", not "English"). Don't try to
"fix" the values that come from the PoolParty metadata,
by converting "en" into "English" for storing
in $scope.vocab.language; let AngularJS do the work
of translating back and forth.