CC-2439 Link checker: don't fail with TypeError in a rare case
During checking of a link, if the following combination of factors
applies:

* the link being checked explicitly specifies a port number (80, 443,
  or whatever)
* the response has status 301/302
* the value of the Location response header is "relative", i.e.,
  doesn't begin with "http(s)"

the link is immediately marked as BROKEN, with the error "Error: must
be str, not int"; the redirect is not followed.

The failure occurs during construction of the "absolute" link to
follow, because of a TypeError: the port was being used in string
concatenation, but the type of the port variable is int, not str.

Cast the value of port to str, so that string concatenation succeeds.